+++ CHANGELOG.md
@@ -5,4 +5,4 @@
 
 ### Features
 
-* add reusable contract drift checks ([120d21f](https://github.com/ExtraToast/api-contract-checks/commit/120d21f96160f543992ef9b508777297690d0557))
+* add reusable contract drift checks ([120d21f](https://github.com/JorisJonkers-dev/api-contract-checks/commit/120d21f96160f543992ef9b508777297690d0557))